[IMP] odoo-print: change action test inheritance
User-story: 1522
Task: 2083
