+++ addons/print/tests/common.py
@@ -1,5 +1,6 @@
 """Printing tests"""
 
+from psycopg2 import IntegrityError
 from contextlib import contextmanager
 from io import BytesIO
 import pathlib
@@ -8,7 +9,7 @@ import sys
 from unittest.mock import patch, Mock, ANY, call
 from lxml import etree
 from odoo.modules.module import get_resource_from_path, get_resource_path
-from odoo.tools import config
+from odoo.tools import config, mute_logger
 from odoo.tools.mimetypes import guess_mimetype
 from odoo.tests import common
 
@@ -175,3 +176,145 @@ class PrinterHttpCase(common.HttpCase):
         # pylint: disable=arguments-differ
         with self.release():
             return super().url_open(*args, **kwargs)
+
+class ActionPrintCase(PrinterCase):
+    """Base printing action tests for all print strategy models.
+
+       This class is not for running tests directly. Instead, classes for
+       running tests must derive from this class. The placement of this
+       class in this module requires that this module is not searched for
+       tests.
+
+       To use this class in a derived class for running tests you must import
+       this class by importing the module which contains this class. If you
+       directly import this class and not the whole module then the unittest
+       framework will find this class and run the tests against an instance
+       of this base class, which is not the intended use. In order to protect
+       against this `strategy_model` is set to None which will cause test
+       failures. (This mechanism requires that 'strategy_id' is mandatory in
+       the data model for `action_model`. If it is not, then the behaviour of
+       running tests in this class is not defined.)
+
+       Derived classes must set `strategy_model` as appropriate. Derived
+       classes are also responsible for maintaining test method numbering.
+       Though unit tests should ideally never assume the order in which they
+       are written, these tests are not unit tests: they are higher level
+       tests executed using the unittest framework. Derived classes must not
+       assume that the order in which tests are executed is unimportant and
+       must maintain test method naming (numbering) such that derived class
+       tests are executed after base class tests. It should be assumed that
+       further test cases will be added to this class.
+    """
+    action_model = 'ir.actions.server'
+    # derived classes MUST set the strategy model to test
+    # None ensures running tests in an instance of this class fails
+    strategy_model = None
+
+    @property
+    def default_report(self):
+        """Return the default report"""
+        return self.env.ref('print.action_report_test_page')
+
+    @property
+    def default_printer(self):
+        """Return the default printer"""
+        return self.env.ref('print.default_printer')
+
+    @classmethod
+    def model_id(cls, model):
+        """Return the model id of `model`."""
+        return cls.env['ir.model']._get_id(model)
+
+    def create_action(self, name, model=True):
+        """Return a new print action with `name` and print strategy `model`.
+           If `model` is True, then use :attr:`strategy_model` instead.
+        """
+        if model is True:
+            model = self.strategy_model
+        return self.env[self.action_model].create({
+            'name': name,
+            'model_id': self.model_id(self.action_model),
+            'state': 'print',
+            'strategy_id': None if model is None else self.model_id(model),
+        })
+
+    def create_strategy(self, name, report, printer,
+                        safety=None, model=True, **kwargs):
+        """Return a new print strategy instance from kwargs overridden by args.
+           If `model` is True, then use :attr:`strategy_model` instead.
+        """
+        if model is True:
+            model = self.strategy_model
+        kwargs.update({
+            'name': name,
+            'report_id': None if report is None else report.id,
+            'printer_id': None if printer is None else printer.id,
+            'safety': safety,
+        })
+        return self.env[model].create(kwargs)
+
+    def action_context(self, obj):
+        """Return an action run context which may select strategies."""
+        return {'active_model': obj._name, 'active_ids': obj.ids}
+
+    def benign_context(self, obj):
+        """Return an action run context which selects no strategies."""
+        return self.action_context(obj)
+
+    def test01_missing_strategy(self):
+        """Test print action must specify strategy_id"""
+        with mute_logger('odoo.sql_db'), self.assertRaises(IntegrityError):
+            self.create_action('missing strategy_id', None)
+
+    def test02_no_strategies(self):
+        """Test print action with no strategies"""
+        action = self.create_action('no strategies')
+        # nothing to be printed when the action is run in the wrong context
+        action.with_context(**self.benign_context(action)).run()
+        self.mock_subprocess.Popen.assert_not_called()
+
+    def test03_missing_report(self):
+        """Test print strategy must specify report_id"""
+        with mute_logger('odoo.sql_db'), self.assertRaises(IntegrityError):
+            self.create_strategy('missing report_id', None, None)
+
+    def test04_missing_printer(self):
+        """Test print strategy may omit printer_id"""
+        self.create_strategy('missing printer_id', self.default_report, None)
+
+    def test05_one_strategy(self):
+        """Test print action with one strategy"""
+        action = self.create_action('one strategy')
+        self.create_strategy('Chesney', self.default_report, None)
+        # nothing to be printed when the action is run in the wrong context
+        action.with_context(**self.benign_context(action)).run()
+        self.mock_subprocess.Popen.assert_not_called()
+        # a single document to be printed when run in the right context
+        printer = self.default_printer
+        action.with_context(**self.action_context(printer)).run()
+        self.assertPrintedLpr('-T', ANY)
+
+    def test06_two_strategies(self):
+        """Test print action with two strategies"""
+        action = self.create_action('two strategies')
+        self.create_strategy('Ant', self.default_report, None)
+        self.create_strategy('Dec', self.default_report, None)
+        # nothing to be printed when the action is run in the wrong context
+        action.with_context(**self.benign_context(action)).run()
+        self.mock_subprocess.Popen.assert_not_called()
+        # two documents to be printed when run in the right context
+        printer = self.default_printer
+        action.with_context(**self.action_context(printer)).run()
+        self.assertPrintedLprMulti(
+            ('-T', ANY),
+            ('-T', ANY),
+        )
+
+    def test07_safety_catch(self):
+        """Test print strategy disabled by safety catch"""
+        action = self.create_action('safety strategy')
+        self.create_strategy('Test safety', self.default_report, None, 'safety')
+        # nothing to be printed when the safety catch is not disabled
+        printer = self.default_printer
+        action.with_context(**self.action_context(printer)).run()
+        self.mock_subprocess.Popen.assert_not_called()
